+++ jgui/jslider.cpp
@@ -104,7 +104,7 @@ void Slider::SetPosition(double i)
 		}
 		
 		jadjustment_type_t t = UNIT_INCREMENT;
-		int diff = _position-_old_position;
+		int diff = (int)(_position-_old_position);
 
 		if (diff > _minimum_tick) {
 			t = BLOCK_INCREMENT;
